Create a base class for the providers
The base class will define the public interface for the provider
classes. This interface acts as the adaptation layer between cloudware
and the provider
